+++ gazebo/gui/LayersWidget_TEST.hh
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2015 Open Source Robotics Foundation
+ * Copyright (C) 2015 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,12 +15,18 @@
  *
 */
 
-#include "gazebo/rendering/RenderEvents.hh"
+#ifndef _GAZEBO_LAYERSWIDGET_TEST_HH_
+#define _GAZEBO_LAYERSWIDGET_TEST_HH_
 
-using namespace gazebo;
-using namespace rendering;
+#include "gazebo/gui/QTestFixture.hh"
 
-event::EventT<void (const std::string &)> Events::createScene;
-event::EventT<void (const std::string &)> Events::removeScene;
-event::EventT<void (const int32_t)> Events::toggleLayer;
-event::EventT<void (const int32_t)> Events::newLayer;
+/// \brief A test class for the Layers widget.
+class LayersWidget_TEST : public QTestFixture
+{
+  Q_OBJECT
+
+  /// \brief Test toggling layers.
+  private slots: void ToggleLayers();
+};
+
+#endif